refactor: common files
