+++ packages/shared/src/features/onboarding/shared/utils.ts
@@ -1,0 +1,20 @@
+import createDOMPurify from 'dompurify';
+
+/**
+ * Sanitizes HTML string and allows only bold tags
+ */
+export const sanitizeMessage = (message: string): string => {
+  // Only run on client-side
+  if (typeof window === 'undefined') {
+    return message;
+  }
+
+  const purify = createDOMPurify(window);
+
+  // Configure DOMPurify to only allow <b> and <strong> tags
+  // and <br> tags for line breaks
+  return purify.sanitize(message, {
+    ALLOWED_TAGS: ['b', 'strong', 'br'],
+    ALLOWED_ATTR: [],
+  });
+};
